backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "NatWest AI Hackathon Backend"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.44.0",
]
